Moved reused functions to fetch_company_from_sirene.py in order to refactor existing js to be replaced with py functions
